# PERF: avoid warnings in IndexEngine.get_loc