--- a/pandas/_libs/index.pyx
+++ b/pandas/_libs/index.pyx
@@ -1,5 +1,3 @@
-import warnings
-
 cimport cython
 
 import numpy as np