--- a/pandas/_libs/index_class_helper.pxi.in
+++ b/pandas/_libs/index_class_helper.pxi.in
@@ -37,8 +37,8 @@ cdef class {{name}}Engine(IndexEngine):
         if not util.is_integer_object(val):
             raise KeyError(val)
     {{else}}
-        if util.is_bool_object(val):
-            # avoid casting to True -> 1.0
+        if not util.is_integer_object(val) and not util.is_float_object(val):
+            # in particular catch bool and avoid casting True -> 1.0
             raise KeyError(val)
     {{endif}}
 
@@ -53,22 +53,15 @@ cdef class {{name}}Engine(IndexEngine):
         self._check_type(val)
 
         values = self.values
-        try:
-            with warnings.catch_warnings():
-                # e.g. if values is float64 and `val` is a str, suppress warning
-                warnings.filterwarnings("ignore", category=FutureWarning)
-                {{if name in {'Float64', 'Float32'} }}
-                if util.is_nan(val):
-                    indexer = np.isnan(values)
-                else:
-                    indexer = values == val
-                {{else}}
-                indexer = values == val
-                {{endif}}
-        except TypeError:
-            # if the equality above returns a bool, cython will raise TypeError
-            #  when trying to cast it to ndarray
-            raise KeyError(val)
+
+        {{if name in {'Float64', 'Float32'} }}
+        if util.is_nan(val):
+            indexer = np.isnan(values)
+        else:
+            indexer = values == val
+        {{else}}
+        indexer = values == val
+        {{endif}}
 
         return indexer